Expressions workaround: add EXPR_EXPORTS when building avt.
The issue is that avtExprNode derives from ExprNode.  In some avt
code, we're handed an ExprNode and we upcast (via dynamic_cast) to
an IntegerConstExpr (for example).  dynamic_cast has to do a
typeinfo lookup, but it can't because ExprNode's typeinfo isn't in
the current shared object.  Since the type lookup fails, the
dynamic_cast fails, and we crash and burn.

Here we add a define to extra-export the common/expr code into
avt.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@7692 18c085ea-50e0-402c-830e-de6fd14e8384