Don't allow overlaps.
We can't composite overlapping images if either has transparency.
I imagine there's a performance gain to be had by allowing some
minor overlaps, but since we can't be sure it will work in all
cases we should disable it entirely for now.

We might think about adding a `has transparency' bool to our
contracts, so we could query that in situations like this one.  Or
we could just fix transparent compositing...

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@8010 18c085ea-50e0-402c-830e-de6fd14e8384