Update GLEW.
This is the generated GLEW 1.5.1 with my and Jianrong Shu's
patches to do runtime loading of the OpenGL library.  This version
requires one to use the alternate entry point glewInitLibrary, and
so it is not recommended that you pull this version for external
use.

We're working on fixing that limitation, and then getting it
pushed upstream.  This will all be resolved before 2.0.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@8016 18c085ea-50e0-402c-830e-de6fd14e8384