Remove initial creation of window 0.
We'll do it dynamically when required.  This has the effect of
delaying the window initialization, decoupling it from the rest of
the engine's initialization.  Thus, when we do get around to
initializing the window, we'll be in a better position for both
detecting and reporting errors, because we can rely on the
previous initialization of our error handling routines.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@9206 18c085ea-50e0-402c-830e-de6fd14e8384