Updated the search logic for multiple libs.
When the user gives multiple libraries, we don't want to bail if
we can't find one.  Now we'll search for a working library and
bail if we find it.  I also downgraded a SEND_ERROR to a WARNING;
the lib_error stuff should still catch configurations that are
real errors.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@10470 18c085ea-50e0-402c-830e-de6fd14e8384