Remove special-cased Mesa code.
Most of these were using a proper runtime-based `if' anyway, so
they should work okay.  However, we don't have any reason to
maintain two implementations anymore, and this removes the
possibility of multiple bits of state disagreeing.

I also fixed some code for detecting point sprites to just use a
GLEW check instead of manually looking for the extension, in a way
which previously assumed features of our Mesa implementation.

git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@11067 18c085ea-50e0-402c-830e-de6fd14e8384