I fixed several compiler warnings where we were passing/setting an int from a float/double. I had nothing better to do waiting for a trunk compile to complete.
The only one that bothers me here is QvisPoincarePlotWindow.C. The code
I 'touched up' was, in theory, generated code. So, why was it wrong? The
generated code (line 1335) was using IntToQString() on a plot's attribute
that was definately of type double. I could have gone two ways; cast arg
to int or call DoubleToQString(). I did the first; cast'd the arg to int.
The attribute's name, singlePlane, suggested it was more a less a boolean.


git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk/src@13908 18c085ea-50e0-402c-830e-de6fd14e8384